requirements.txt
asgiref==3.3.1
astroid==2.4.2
beautifulsoup4==4.9.3
cairocffi==1.2.0
CairoSVG==2.5.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorclass==2.2.0
cssselect2==0.4.1
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.6
django-admin-list-filter-dropdown==1.0.3
django-appconf==1.0.4
django-bootstrap4==2.3.1
django-braces==1.14.0
django-crispy-forms==1.11.0
django-import-export==2.5.0
django-reverse-admin==2.9.4
django-select2==7.6.1
django-tagify==0.20
django-weasyprint==1.0.2
docopt==0.6.2
et-xmlfile==1.0.1
furl==2.1.0
gunicorn==20.0.4
html5lib==1.1
idna==2.10
importlib-metadata==2.1.1
isort==5.7.0
jdcal==1.4.1
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.6
orderedmultidict==1.0.1
packaging==20.9
Pillow==8.1.0
pip-upgrader==1.4.15
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.7
Pyphen==0.10.0
python-dotenv==0.15.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
simplejson==3.17.2
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
tablib==3.0.0
terminaltables==3.1.0
tinycss2==1.1.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
WeasyPrint==52.2
webencodings==0.5.1
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.4.0
./custom_libs/django-fdigital-sso-0.1.tar.gz